#54 - Unified naming conventions
